requirements.txt
absl-py==2.4.0
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
async-timeout==5.0.1
attrs==26.1.0
bleach==6.3.0
blinker==1.9.0
cachetools==7.0.6
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
clearml==2.1.5
click==8.3.3
cloudpickle==3.1.2
contourpy==1.3.2
cryptography==46.0.7
cuda-bindings==13.2.0
cuda-pathfinder==1.5.3
cuda-toolkit==13.0.2
cycler==0.12.1
databricks-sdk==0.67.0
defusedxml==0.7.1
docker==7.1.0
exceptiongroup==1.3.1
fastapi==0.136.1
faster-coco-eval==1.7.2
filelock==3.29.0
filetype==1.2.0
Flask==3.1.3
flask-cors==6.0.2
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.3.0
furl==2.1.4
gitdb==4.0.12
GitPython==3.1.47
google-auth==2.49.2
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
greenlet==3.4.0
grpcio==1.80.0
gunicorn==25.3.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huey==2.6.0
huggingface_hub==1.12.0
idna==3.7
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kaggle==1.7.4.5
kagglehub==1.0.0
kagglesdk==0.1.21
kiwisolver==1.5.0
lightning-utilities==0.15.3
Mako==1.3.11
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.9
mdurl==0.1.2
mlflow==3.11.1
mlflow-skinny==3.11.1
mlflow-tracing==3.11.1
mpmath==1.3.0
multidict==6.7.1
networkx==3.4.2
numpy==2.2.6
nvidia-cublas==13.1.0.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti==13.0.85
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime==13.0.96
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufft-cu11==10.9.0.58
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver==12.0.4.66
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse==12.6.3.3
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
nvidia-nvtx-cu11==11.8.86
opencv-python==4.13.0.92
opencv-python-headless==4.10.0.84
opentelemetry-api==1.41.1
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.41.1
opentelemetry-semantic-conventions==0.62b1
orderedmultidict==1.0.2
packaging==26.0
panda==0.3.1
pandas==2.3.3
pathlib2==2.3.7.post1
peft==0.19.1
pi_heif==1.3.0
pillow==12.2.0
pillow-avif-plugin==1.5.5
platformdirs==4.9.6
prettytable==3.17.0
propcache==0.4.1
protobuf==3.20.3
psutil==7.2.2
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycocotools==2.0.11
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
pyDeprecate==0.5.0
Pygments==2.20.0
PyJWT==2.12.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-slugify==8.0.4
pytorch-lightning==2.6.1
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
rf100vl==1.1.0
rfdetr==1.6.5
rich==15.0.0
roboflow==1.3.5
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentry-sdk==2.58.0
shellingham==1.5.4
simsimd==6.5.16
six==1.17.0
skops==0.14.0
smmap==5.0.3
SQLAlchemy==2.0.49
sqlparse==0.5.5
starlette==1.0.0
stringzilla==4.6.0
supervision==0.27.0.post2
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
text-unidecode==1.3
threadpoolctl==3.6.0
tokenizers==0.22.2
tomli==2.4.1
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchmetrics==1.9.0
torchvision==0.22.1+cu118
tqdm==4.67.3
transformers==5.6.2
triton==3.3.1
typer==0.25.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
uvicorn==0.46.0
wandb==0.25.0
wcwidth==0.6.0
webencodings==0.5.1
Werkzeug==3.1.8
yarl==1.23.0
zipp==3.23.1
